server/verify_chat.py
import requests
import sys

# Constants
BASE_URL = "http://localhost:5050/api"
EMAIL = "test@example.com" # Replace with valid user if known, or I'll need to login
PASSWORD = "password123"

def verify_persistence():
    try:
        # 1. Login
        # I need a valid user. I'll check the database first to find one or create one.
        # Since I can't interactively ask for credentials easily in a script without stalling,
        # I will use the python shell tool to inspect the DB directly first.
        pass
    except Exception as e:
        print(e)
