server/requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-mail
psycopg2-binary
python-dotenv
bcrypt
PyJWT
llama-index
llama-index-llms-groq
llama-index-embeddings-huggingface
PyMuPDF
sentence-transformers
torch
